Fixed bug when file with precalculated tiles was deleted

+++ src/main/kotlin/com/github/kornilova_l/algorithm_synthesis/grid2D/tiles/TileGenerator.kt
@@ -44,7 +44,9 @@ abstract class TileGenerator<T : BinaryTile>(private val finalN: Int, private va
         var currentSize = FileNameCreator.getIntParameter(currentTilesFile.name, "size")!!
         while (currentM < finalM || currentN < finalN) {
             val pair = expandTileSet(currentTilesFile, currentN, currentM)
-            currentTilesFile.delete()
+            if (currentTilesFile.absoluteFile.parent == tempDir.name) { // if inside temp file
+                currentTilesFile.delete()
+            }
             currentTilesFile = pair.first
             currentSize = pair.second
             currentN = FileNameCreator.getIntParameter(currentTilesFile.name, "n")!!
@@ -53,8 +55,12 @@ abstract class TileGenerator<T : BinaryTile>(private val finalN: Int, private va
         size = currentSize
         file = Paths.get(outputDir.toString(),
                 tilesFileNameCreator.getFileName(finalN, finalM, size)).toFile()
-        currentTilesFile.renameTo(file)
-        println("output file: $file")
+        val res = currentTilesFile.renameTo(file.absoluteFile)
+        if (res) {
+            println("Output file: $file")
+        } else {
+            System.err.println("Cannot rename file to $file")
+        }
         Util.deleteDir(tempDir.toPath())
     }
